Rename `Rejections` to `StandardRejections`
