+++ server/src/main/proto/spine/server/entity/standard_rejections.proto
@@ -26,7 +26,7 @@ import "spine/options.proto";
 option (type_url_prefix) = "type.spine.io";
 option (SPI_all) = true;
 option java_package = "io.spine.server.entity.rejection";
-// Keep all the rejections under the same outer class called `Rejections`.
+// Keep all the rejections under the same outer class called `StandardRejections`.
 option java_generate_equals_and_hash = true;
 
 import "google/protobuf/any.proto";
