+++ server/src/test/java/io/spine/server/procman/ProcessManagerRepositoryShould.java
@@ -41,8 +41,8 @@ import io.spine.server.BoundedContext;
 import io.spine.server.entity.RecordBasedRepository;
 import io.spine.server.entity.RecordBasedRepositoryShould;
 import io.spine.server.entity.given.Given;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyArchived;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyDeleted;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyArchived;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyDeleted;
 import io.spine.server.event.EventSubscriber;
 import io.spine.server.procman.given.ProcessManagerRepositoryTestEnv;
 import io.spine.server.procman.given.ProcessManagerRepositoryTestEnv.TestProcessManager;
