+++ server/src/test/java/io/spine/server/procman/given/ProcessManagerRepositoryTestEnv.java
@@ -29,8 +29,8 @@ import io.spine.core.EventContext;
 import io.spine.core.React;
 import io.spine.server.command.Assign;
 import io.spine.server.entity.TestEntityWithStringColumn;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyArchived;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyDeleted;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyArchived;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyDeleted;
 import io.spine.server.procman.CommandRouted;
 import io.spine.server.procman.ProcessManager;
 import io.spine.server.procman.ProcessManagerRepository;
