+++ server/src/test/java/io/spine/server/route/RejectionRoutingShould.java
@@ -27,8 +27,8 @@ import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import io.spine.Identifier;
 import io.spine.core.RejectionContext;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyArchived;
-import io.spine.server.entity.rejection.Rejections.EntityAlreadyDeleted;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyArchived;
+import io.spine.server.entity.rejection.StandardRejections.EntityAlreadyDeleted;
 import org.junit.Before;
 import org.junit.Test;
